+++ src/paco/application/ec2lm_commands.py
@@ -21,6 +21,11 @@ user_data_script = {
 	},
 	'install_aws_cli': {
 		'amazon': '', # AWS is installed by default on Amazon linux
+        'amazon_ecs': """yum -y install unzip
+curl "https://awscli.amazonaws.com/awscli-exe-linux-x86_64.zip" -o "awscliv2.zip"
+unzip awscliv2.zip
+./aws/install --bin-dir /usr/bin
+""",
 		'ubuntu': """apt-get update
 apt-get -y install python-pip
 pip install awscli